+++ test/EasyPost/Fixture.php
@@ -122,7 +122,7 @@ class Fixture
     // USPS only does "next-day" pickups including Saturday but not Sunday or Holidays.
     public static function basicPickup()
     {
-        $pickupDate = '2022-12-09';
+        $pickupDate = '2023-11-24';
 
         $pickupData = self::readFixtureData()['pickups']['basic'];
         $pickupData['min_datetime'] = $pickupDate;
@@ -179,6 +179,6 @@ class Fixture
 
     public static function plannedShipDate()
     {
-        return '2023-04-28';
+        return '2023-11-24';
     }
 }